python/requirements.txt
watchdog==0.9.0
pycrypto==2.6.1
